Expose common functions used by ManifestReader
